Fix a couple of issues with the CMake script
Summary:
The first issue is that I was missing a `G` in `LIBGFLAGS_INCLUDE_DIR` meaning that it always claimed it couldn't find gflags even though it did, and worked just fine anyways.
The second issue is that it was passing `/permissive-` to MSVC 2015 builds, even though MSVC 2015 doesn't support the permissive flag.

Reviewed By: yfeldblum

Differential Revision: D4988813

fbshipit-source-id: f1ea009226baee54032409ce7add3c41f1fe7a19